Remove unnecesary controller and routing
